Created ClipboardUtil to generate copy-to-clipboard buttons
