+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/utility/GeneralUtils.kt
@@ -2,8 +2,15 @@ package com.github.dedis.popstellar.utility
 
 import android.app.Activity
 import android.app.Application
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
 import android.os.Bundle
+import android.view.View
+import android.widget.TextView
+import android.widget.Toast
 import androidx.lifecycle.Lifecycle
+import com.github.dedis.popstellar.R
 import com.github.dedis.popstellar.model.objects.security.Base64URLData
 import io.github.novacrypto.bip39.MnemonicGenerator
 import io.github.novacrypto.bip39.wordlists.English
@@ -64,6 +71,30 @@ object GeneralUtils {
       }
     }
   }
+
+  /**
+   * This class is a utility to setup a copy button that copies the text of a TextView to the
+   * clipboard.
+   *
+   * @param context the context of the application
+   */
+  class ClipboardUtil(val context : Context) {
+
+    fun setupCopyButton(button: View, textView: TextView, label: String) {
+      button.setOnClickListener {
+        val text = textView.text.toString()
+        copyTextToClipboard(label, text)
+        Toast.makeText(context, R.string.successful_copy, Toast.LENGTH_SHORT).show()
+      }
+    }
+
+    private fun copyTextToClipboard(label: String, content: String) {
+      val clipboard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+      val clip = ClipData.newPlainText(label, content)
+      clipboard.setPrimaryClip(clip)
+    }
+  }
+
 
   /**
    * This function converts a base64 string into some mnemonic words.
